feat: api 타입스크립트 파일로 변경
